requirements.txt
fastapi
uvicorn
langchain
sentence-transformers
faiss-cpu
pymupdf
transformers
accelerate
python-multipart